Move queue commands to subcommand and standardise naming
